backend/requirements.txt
Django==4.2.16
pillow==10.4.0
djangorestframework==3.15.2
djoser==2.1.0
django-filter==23.1
python-dotenv==1.0.1
flake8==7.1.1